Additional manifest submission tests and migration of tests
