+++ tests/test_api.py
@@ -1096,217 +1096,6 @@ class TestManifestOperation:
             assert isinstance(response_path, str)
             assert response_path.endswith(".csv")
 
-    @pytest.mark.synapse_credentials_needed
-    @pytest.mark.submission
-    def test_submit_manifest_table_and_file_replace(
-        self,
-        client: FlaskClient,
-        request_headers: Dict[str, str],
-        test_manifest_submit: str,
-    ) -> None:
-        """Testing submit manifest in a csv format as a table and a file. Only replace the table"""
-        params = {
-            "schema_url": DATA_MODEL_JSON_LD,
-            "data_type": "Biospecimen",
-            "restrict_rules": False,
-            "hide_blanks": False,
-            "manifest_record_type": "table_and_file",
-            "asset_view": "syn51514344",
-            "dataset_id": "syn51514345",
-            "table_manipulation": "replace",
-            "data_model_labels": "class_label",
-            "table_column_names": "class_label",
-        }
-
-        response_csv = client.post(
-            "http://localhost:3001/v1/model/submit",
-            query_string=params,
-            data={"file_name": (open(test_manifest_submit, "rb"), "test.csv")},
-            headers=request_headers,
-        )
-        assert response_csv.status_code == 200
-
-    @pytest.mark.synapse_credentials_needed
-    @pytest.mark.submission
-    @pytest.mark.parametrize(
-        "data_type, manifest_path_fixture",
-        [
-            ("Biospecimen", "test_manifest_submit"),
-            ("MockComponent", "valid_test_manifest_csv"),
-        ],
-    )
-    def test_submit_manifest_file_only_replace(
-        self,
-        helpers,
-        client: FlaskClient,
-        request_headers: Dict[str, str],
-        data_type: str,
-        manifest_path_fixture: str,
-        request: pytest.FixtureRequest,
-    ) -> None:
-        """Testing submit manifest in a csv format as a file"""
-        params = {
-            "schema_url": DATA_MODEL_JSON_LD,
-            "data_type": data_type,
-            "restrict_rules": False,
-            "manifest_record_type": "file_only",
-            "table_manipulation": "replace",
-            "data_model_labels": "class_label",
-            "table_column_names": "class_label",
-        }
-
-        if data_type == "Biospecimen":
-            specific_params = {
-                "asset_view": "syn51514344",
-                "dataset_id": "syn51514345",
-            }
-
-        elif data_type == "MockComponent":
-            python_version = helpers.get_python_version()
-
-            if python_version == "3.10":
-                dataset_id = "syn52656106"
-            elif python_version == "3.9":
-                dataset_id = "syn52656104"
-
-            specific_params = {
-                "asset_view": "syn23643253",
-                "dataset_id": dataset_id,
-                "project_scope": ["syn54126707"],
-            }
-
-        params.update(specific_params)
-
-        manifest_path = request.getfixturevalue(manifest_path_fixture)
-        response_csv = client.post(
-            "http://localhost:3001/v1/model/submit",
-            query_string=params,
-            data={"file_name": (open(manifest_path, "rb"), "test.csv")},
-            headers=request_headers,
-        )
-        assert response_csv.status_code == 200
-
-    @pytest.mark.synapse_credentials_needed
-    @pytest.mark.submission
-    def test_submit_manifest_json_str_replace(
-        self, client: FlaskClient, request_headers: Dict[str, str]
-    ) -> None:
-        """Submit json str as a file"""
-        json_str = '[{"Sample ID": 123, "Patient ID": 1,"Tissue Status": "Healthy","Component": "Biospecimen"}]'
-        params = {
-            "schema_url": DATA_MODEL_JSON_LD,
-            "data_type": "Biospecimen",
-            "json_str": json_str,
-            "restrict_rules": False,
-            "manifest_record_type": "file_only",
-            "asset_view": "syn51514344",
-            "dataset_id": "syn51514345",
-            "table_manipulation": "replace",
-            "data_model_labels": "class_label",
-            "table_column_names": "class_label",
-        }
-        params["json_str"] = json_str
-        response = client.post(
-            "http://localhost:3001/v1/model/submit",
-            query_string=params,
-            data={"file_name": ""},
-            headers=request_headers,
-        )
-        assert response.status_code == 200
-
-    @pytest.mark.synapse_credentials_needed
-    @pytest.mark.submission
-    def test_submit_manifest_w_file_and_entities(
-        self,
-        client: FlaskClient,
-        request_headers: Dict[str, str],
-        test_manifest_submit: str,
-    ) -> None:
-        params = {
-            "schema_url": DATA_MODEL_JSON_LD,
-            "data_type": "Biospecimen",
-            "restrict_rules": False,
-            "manifest_record_type": "file_and_entities",
-            "asset_view": "syn51514501",
-            "dataset_id": "syn51514523",
-            "table_manipulation": "replace",
-            "data_model_labels": "class_label",
-            "table_column_names": "class_label",
-            "annotation_keys": "class_label",
-        }
-
-        # test uploading a csv file
-        response_csv = client.post(
-            "http://localhost:3001/v1/model/submit",
-            query_string=params,
-            data={"file_name": (open(test_manifest_submit, "rb"), "test.csv")},
-            headers=request_headers,
-        )
-        assert response_csv.status_code == 200
-
-    @pytest.mark.synapse_credentials_needed
-    @pytest.mark.submission
-    def test_submit_manifest_table_and_file_upsert(
-        self,
-        client: FlaskClient,
-        request_headers: Dict[str, str],
-        test_upsert_manifest_csv: str,
-    ) -> None:
-        params = {
-            "schema_url": DATA_MODEL_JSON_LD,
-            "data_type": "MockRDB",
-            "restrict_rules": False,
-            "manifest_record_type": "table_and_file",
-            "asset_view": "syn51514557",
-            "dataset_id": "syn51514551",
-            "table_manipulation": "upsert",
-            "data_model_labels": "class_label",
-            # have to set table_column_names to display_name to ensure upsert feature works
-            "table_column_names": "display_name",
-        }
-
-        # test uploading a csv file
-        response_csv = client.post(
-            "http://localhost:3001/v1/model/submit",
-            query_string=params,
-            data={"file_name": (open(test_upsert_manifest_csv, "rb"), "test.csv")},
-            headers=request_headers,
-        )
-        assert response_csv.status_code == 200
-
-    @pytest.mark.synapse_credentials_needed
-    @pytest.mark.submission
-    def test_submit_and_validate_filebased_manifest(
-        self,
-        client: FlaskClient,
-        request_headers: Dict[str, str],
-        valid_filename_manifest_csv: str,
-    ) -> None:
-        # GIVEN the appropriate upload parameters
-        params = {
-            "schema_url": DATA_MODEL_JSON_LD,
-            "data_type": "MockFilename",
-            "restrict_rules": False,
-            "manifest_record_type": "file_and_entities",
-            "asset_view": "syn23643253",
-            "dataset_id": "syn62822337",
-            "project_scope": "syn23643250",
-            "dataset_scope": "syn62822337",
-            "data_model_labels": "class_label",
-            "table_column_names": "class_label",
-        }
-
-        # WHEN a filebased manifest is validated with the filenameExists rule and uploaded
-        response_csv = client.post(
-            "http://localhost:3001/v1/model/submit",
-            query_string=params,
-            data={"file_name": (open(valid_filename_manifest_csv, "rb"), "test.csv")},
-            headers=request_headers,
-        )
-
-        # THEN the validation and submission should be successful
-        assert response_csv.status_code == 200
-
 
 @pytest.mark.schematic_api
 class TestSchemaVisualization:
